Prefer unused labels to creating a new one
Before this change, loops and conditionals unconditionally create a new
label for the start of the loop. After it, a label that was just added
will be used instead of creating a new one. This can significantly cut
down on the number of label equates (`LBL1 EQU LBL2`) and makes the
assembly simpler.