Associate labels during code generation
MERLIN and a2asm don't support label-only lines. Before this change,
labels were added as operation-less instructions and then associated
with an operation when written out. With this change, the association is
handled during code generation. This will allow for simpler optimization
code and (apparently) is less error prone as I discovered several errors
in the test code that I hadn't noticed before.